aaaaaaaaaaaaaaaaaaaaaaaaaaaa
Now there is only one thing. Only darkness. I was fighting in huge battels against the frightening beasts like CSS styles. I was things that I won. But now the world is on the edge of its destruction.
The last battle is near. We'll see if the sun ever shines again